[CFF] Test nested outline budget charges
Exercise CFF work accounting through the public draw-budget callbacks.
Use lower bounds above each parent charstring's flat operator cost so the
CFF1 assertion covers both seac children and the CFF2 assertion covers
proportional blend work.

Keep the interpreter's local cost type consistent with the unsigned spend
interface.

Tests: meson test -C build (275/275 passed)

+++ src/hb-cff-interp-cs-common.hh
@@ -893,8 +893,8 @@ struct cs_interpreter_t : interpreter_t<ENV>
     unsigned max_ops = HB_CFF_MAX_OPS;
     for (;;) {
       op_code_t op = SUPER::env.fetch_op ();
-      int64_t cost = op == OpCode_callsubr || op == OpCode_callgsubr ?
-		     HB_BUDGET_8 : HB_BUDGET_1;
+      unsigned int cost = op == OpCode_callsubr || op == OpCode_callgsubr ?
+			  HB_BUDGET_8 : HB_BUDGET_1;
       if (unlikely (!SUPER::env.spend_budget (cost)))
       {
 	SUPER::env.set_error ();
